ARPO/requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiolimiter==1.2.1
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
arxiv==2.2.0
astor==0.8.1
debugpy>=1.8.0
astral==3.2
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blake3==1.0.4
boggle-solver==1.1
boto3==1.38.44
botocore==1.38.44
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chembl-webresource-client==0.10.9
chess==1.11.2
click==8.1.8
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.6
compressed-tensors==0.9.3
contourpy==1.3.2
coolprop==6.8.0
cryptography==45.0.4
cupy-cuda12x==13.4.1
curl_cffi==0.11.4
cycler==0.12.1
datasets==3.5.0
dateparser==1.2.1
decorator==5.2.1
deepspeed==0.16.7
Deprecated==1.2.18
depyf==0.18.0
deuces==0.2.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
easydict==1.13
einops==0.8.1
email_validator==2.2.0
ephem==4.2
equivalence==0.1
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
flash_attn==2.7.4.post1
flexcache==0.3
flexparser==0.4
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2024.6.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gguf==0.16.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
gradio==5.25.0
gradio_client==1.8.0
greenlet==3.2.1
groovy==0.1.2
grpcio==1.72.0
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.0.0
interegular==0.3.3
jellyfish==1.2.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langid==1.1.6
lark==1.2.2
lazy_loader==0.4
librosa==0.11.0
llguidance==0.7.19
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mistral_common==1.5.4
modelscope==1.26.0
morsepy==0.3.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
multitasking==0.0.11
mwclient==0.10.1
mwparserfromhell==0.6
nest-asyncio==1.6.0
netcat==0.0.1a0
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.3.1
omegaconf==2.3.0
openai==1.76.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
openreview-py==1.51.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.3
orjson==3.10.16
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
patsy==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.1
peft==0.15.1
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.7
playwright==1.51.0
pooch==1.8.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
PubChemPy==1.0.4
PuLP==3.2.1
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybaseball==2.2.7
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
PyNaCl==1.5.0
pyparsing==3.2.3
pypdf==5.7.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-whois==0.9.5
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ratelim==0.1.6
ray==2.46.0
rdkit-pypi==2022.9.5
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.3
rouge-chinese==1.0.3
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.7
s3transfer==0.13.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.5
sgmllib3k==1.0.0
shellingham==1.5.4
shelved-cache==0.3.1
shtab==1.7.2
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spotipy==2.25.1
sse-starlette==2.3.3
starlette==0.46.2
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.8.10
tensordict==0.6.2
termcolor==3.0.1
thermostate==2.0.0.post1
threadpoolctl==3.6.0
tiktoken==0.9.0
tld==0.13.1
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0+cu124
torchaudio==2.6.0
torchdata==0.11.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
treys==0.1.8
trimesh==4.6.13
triton==3.2.0
trl==0.9.6
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
url-normalize==2.2.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
verl==0.4.0
virtualenv==20.30.0
vllm==0.8.4
wandb==0.19.10
watchfiles==1.0.5
waybackpy==3.0.6
wbdata==1.0.0
wbgapi==1.1.2
websockets==15.0.1
wikipedia==1.4.0
Wikipedia-API==0.5.4
wikipedia-histories==1.1.0
wordcloud==1.9.4
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.0
yfinance==0.2.65
zipp==3.21.0
